Settings Menu POC; Logoff button changed.
Scrolling settings menu, use up and down to alternate.
Logoff changed to select
If you hold B for slightly too long on the settings menu the mainmenu also registers it and logs you off so I just changed it until I can figure out how to wait for the buttons to not be pressed.
Anyone know how to make it wait for the button to be let go of?