Big Commit.
Technically it's not much different than the last in terms of functionality but I converted it to C++

This is also considered Alpha 1.9 by me.

I haven't noticed any bugs (that weren't already in the last commit) but if anyone finds anything please post it in the issue's section so I can begin working on it or taking suggestions right away.