Minor version display changes + menu upDown / leftRight
Alpha and _1.11.3_ are now seperate from each other.
I only did this for coloring.
Also I experimented with some easter eggs in the credits, but failed. Feel free to check the code out and see what I was doing.

Now if you are in a menu and reached the top it won't re-load the page (thus, no more flicker for no reason.) [Not implemented on all menu's yet]